refactor : 추상 클래스 도입
